translation unit for generating numbers
